+++ natural-language-understanding/src/main/java/com/ibm/watson/developer_cloud/natural_language_understanding/v1/model/EntitiesResult.java
@@ -14,7 +14,6 @@ package com.ibm.watson.developer_cloud.natural_language_understanding.v1.model;
 
 import com.ibm.watson.developer_cloud.natural_language_understanding.v1.model.EmotionScores;
 import com.ibm.watson.developer_cloud.natural_language_understanding.v1.model.FeatureSentimentResults;
-import java.math.BigDecimal;
 
 import com.ibm.watson.developer_cloud.natural_language_understanding.v1.NaturalLanguageUnderstanding;
 import com.ibm.watson.developer_cloud.service.model.GenericModel;
@@ -27,7 +26,7 @@ public class EntitiesResult extends GenericModel {
   /** Entity type. */
   private String type;
   /** Relevance score from 0 to 1. Higher values indicate greater relevance. */
-  private BigDecimal relevance;
+  private Long relevance;
   /** How many times the entity was mentioned in the text. */
   private Integer count;
   /** The name of the entity. */
@@ -47,7 +46,7 @@ public class EntitiesResult extends GenericModel {
    * @param emotion Emotion analysis results for the entity, enabled with the \&quot;emotion\&quot; option.
    * @param sentiment Sentiment analysis results for the entity, enabled with the \&quot;sentiment\&quot; option.
    */
-  public EntitiesResult(final String type, final BigDecimal relevance, final Integer count, final String text, final EmotionScores emotion, final FeatureSentimentResults sentiment) {
+  public EntitiesResult(final String type, final Long relevance, final Integer count, final String text, final EmotionScores emotion, final FeatureSentimentResults sentiment) {
     this.type = type;
     this.relevance = relevance;
     this.count = count;
@@ -70,7 +69,7 @@ public class EntitiesResult extends GenericModel {
    *
    * @return the relevance
    */
-  public BigDecimal getRelevance() {
+  public Long getRelevance() {
     return relevance;
   }
 
@@ -124,7 +123,7 @@ public class EntitiesResult extends GenericModel {
    *
    * @param relevance the new relevance
    */
-  public void setRelevance(final BigDecimal relevance) {
+  public void setRelevance(final Long relevance) {
     this.relevance = relevance;
   }
